Implement basic layout and page flow.
* keep root path page but move wordlist onto a /wordlist route and page
