+++ components/Header.js
@@ -1,0 +1,14 @@
+const headerStyle = {
+  backgroundColor: 'blue',
+  color: 'white',
+  width: '100%',
+  height: '50px'
+};
+
+const Header = () => (
+  <div className='Header' style={headerStyle}>
+    HEADER
+  </div>
+);
+
+export default Header;
