+++ components/NavBar.js
@@ -1,0 +1,14 @@
+const navBarStyle = {
+  backgroundColor: 'red',
+  color: 'white',
+  width: '100%',
+  height: '60px'
+};
+
+const NavBar = () => (
+  <div className='NavBar' style={navBarStyle}>
+    NAVBAR
+  </div>
+);
+
+export default NavBar;
